+++ anch-crypto/inc/crypto/cipher/pcbc.hpp
@@ -95,7 +95,7 @@ namespace anch {
 				      std::streamsize nbRead,
 				      std::array<uint8_t,Cipher::getBlockSize()>& output,
 				      uint32_t, Cipher& cipher) override {
-	if(nbRead != Cipher::getBlockSize()) {
+	if(static_cast<std::size_t>(nbRead) != Cipher::getBlockSize()) {
 	  Padding::pad(input.data(), nbRead, Cipher::getBlockSize());
 	}
 	std::array<uint8_t,Cipher::getBlockSize()> data;
@@ -127,7 +127,7 @@ namespace anch {
 					bool lastBlock,
 					std::array<uint8_t,Cipher::getBlockSize()>& output,
 					uint32_t, Cipher& cipher) override {
-	if(lastBlock && nbRead != Cipher::getBlockSize()) {
+	if(lastBlock && static_cast<std::size_t>(nbRead) != Cipher::getBlockSize()) {
 	  throw InvalidBlockException("Invalid block size");
 	}
 	std::array<uint8_t,Cipher::getBlockSize()> data;
